celery_app.py
from celery import Celery
from dotenv import load_dotenv
import os

load_dotenv()

celery = Celery(
    'sounds-right',
    broker=os.getenv('REDIS_URL', 'redis://localhost:6379/0'),
    backend=os.getenv('REDIS_URL', 'redis://localhost:6379/0')
)

celery.conf.update(
    task_serializer='json',
    accept_content=['json'],
    result_serializer='json',
    timezone='UTC',
    enable_utc=True,
    imports=('tasks',)
) 