requirements.txt
celery>=5.3.6
redis>=5.0.1
flask>=3.0.0
pymongo>=4.6.0
python-dotenv>=1.0.0
openai-whisper>=20231117
google-cloud-storage>=2.14.0
pyjwt>=2.8.0
pydub>=0.25.1
numpy>=1.26.0
markupsafe>=2.1.5
werkzeug>=3.0.0
flake8>=7.0.0
google-cloud-spanner>=4.0.0
